Add common entity footer
