+++ src/main/java/jp/mydns/projectk/safi/entity/UserEntity.java
@@ -30,14 +30,10 @@ import jakarta.persistence.Cacheable;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
-import jakarta.persistence.OptimisticLockException;
 import jakarta.persistence.Table;
-import jakarta.persistence.Version;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.PositiveOrZero;
 import jakarta.validation.constraints.Size;
-import java.io.Serializable;
 import java.time.LocalDateTime;
 import java.time.OffsetDateTime;
 import java.util.Objects;
@@ -56,7 +52,7 @@ import jp.mydns.projectk.safi.value.JsonObjectValue;
 @Entity
 @Cacheable(false)
 @Table(name = "t_user")
-public class UserEntity implements Serializable {
+public class UserEntity extends CommonEntity {
 
     private static final long serialVersionUID = 2711050439353117979L;
 
@@ -89,30 +85,6 @@ public class UserEntity implements Serializable {
     @Basic(optional = false)
     @Column(name = "digest", nullable = false, length = 128)
     private String digest;
-
-    @Column(name = "note")
-    private String note;
-
-    @Column(name = "version", nullable = false)
-    private int version;
-
-    @Column(name = "reg_ts", updatable = false)
-    private LocalDateTime registerTime;
-
-    @Column(name = "reg_id", updatable = false, length = 250)
-    private String registerAccountId;
-
-    @Column(name = "reg_ap", updatable = false, length = 250)
-    private String registerProcessName;
-
-    @Column(name = "upd_ts", insertable = false)
-    private LocalDateTime updateTime;
-
-    @Column(name = "upd_id", insertable = false, length = 250)
-    private String updateAccountId;
-
-    @Column(name = "upd_ap", insertable = false, length = 250)
-    private String updateProcessName;
 
     /**
      * Get user id.
@@ -304,179 +276,6 @@ public class UserEntity implements Serializable {
      */
     public void setDigest(String digest) {
         this.digest = digest;
-    }
-
-    /**
-     * Get a note for this entity. This value is only used to record notes about the data records represented by the
-     * entity and is never used to process.
-     *
-     * @return note. It may be {@code null}.
-     * @since 3.0.0
-     */
-    public String getNote() {
-        return note;
-    }
-
-    /**
-     * Set a note for this entity.
-     *
-     * @param note note. It can be set {@code null}.
-     * @since 3.0.0
-     */
-    public void setNote(String note) {
-        this.note = note;
-    }
-
-    /**
-     * Get entity version. This value used for optimistic locking. Will be thrown {@link OptimisticLockException} when
-     * inserted or updated when the configured version number is not equal to that of the database. The version of the
-     * entity before persistence is 0.
-     *
-     * @return entity version
-     * @since 3.0.0
-     */
-    @PositiveOrZero
-    @Version
-    public int getVersion() {
-        return version;
-    }
-
-    /**
-     * Set entity version.
-     *
-     * @param version entity version
-     * @since 3.0.0
-     */
-    public void setVersion(int version) {
-        this.version = version;
-    }
-
-    /**
-     * Get time the entity was persisted.
-     *
-     * @return persisted time. It time zone is UTC. It may be {@code null}.
-     * @since 3.0.0
-     */
-    @TimeRange
-    @TimeAccuracy
-    public LocalDateTime getRegisterTime() {
-        return registerTime;
-    }
-
-    /**
-     * Set time the entity was persisted.
-     *
-     * @param registerTime persisted time. It time zone is UTC. It can be set {@code null}.
-     * @since 3.0.0
-     */
-    public void setRegisterTime(LocalDateTime registerTime) {
-        this.registerTime = registerTime;
-    }
-
-    /**
-     * Get id of the account who made the entity persistent.
-     *
-     * @return persisted account id. It may be {@code null}.
-     * @since 3.0.0
-     */
-    @Size(max = 250)
-    public String getRegisterAccountId() {
-        return registerAccountId;
-    }
-
-    /**
-     * Set id of the account who made the entity persistent.
-     *
-     * @param registerAccountId persisted account id. It can be set {@code null}.
-     * @since 3.0.0
-     */
-    public void setRegisterAccountId(String registerAccountId) {
-        this.registerAccountId = registerAccountId;
-    }
-
-    /**
-     * Get name of the process who made the entity persistent.
-     *
-     * @return persisted process name. It may be {@code null}.
-     * @since 3.0.0
-     */
-    @Size(max = 250)
-    public String getRegisterProcessName() {
-        return registerProcessName;
-    }
-
-    /**
-     * Set name of the process who made the entity persistent.
-     *
-     * @param registerProcessName persisted process name. It can be set {@code null}.
-     * @since 3.0.0
-     */
-    public void setRegisterProcessName(String registerProcessName) {
-        this.registerProcessName = registerProcessName;
-    }
-
-    /**
-     * Get time the entity was last updated.
-     *
-     * @return last updated time. It time zone is UTC. It may be {@code null}.
-     * @since 3.0.0
-     */
-    @TimeRange
-    @TimeAccuracy
-    public LocalDateTime getUpdateTime() {
-        return updateTime;
-    }
-
-    /**
-     * Set time the entity was last updated.
-     *
-     * @param updateTime last updated time. It time zone is UTC. It can be set {@code null}.
-     * @since 3.0.0
-     */
-    public void setUpdateTime(LocalDateTime updateTime) {
-        this.updateTime = updateTime;
-    }
-
-    /**
-     * Get id of the account who made the entity last updated.
-     *
-     * @return last updated account id. It may be {@code null}.
-     * @since 3.0.0
-     */
-    @Size(max = 250)
-    public String getUpdateAccountId() {
-        return updateAccountId;
-    }
-
-    /**
-     * Set id of the account who made the entity last updated.
-     *
-     * @param updateAccountId last updated account id. It can be set {@code null}.
-     * @since 3.0.0
-     */
-    public void setUpdateAccountId(String updateAccountId) {
-        this.updateAccountId = updateAccountId;
-    }
-
-    /**
-     * Get name of the process who made the entity last updated.
-     *
-     * @return last updated process name. It may be {@code null}.
-     * @since 3.0.0
-     */
-    @Size(max = 250)
-    public String getUpdateProcessName() {
-        return updateProcessName;
-    }
-
-    /**
-     * Set name of the process who made the entity last updated.
-     *
-     * @param updateProcessName last updated process name. It can be set {@code null}.
-     * @since 3.0.0
-     */
-    public void setUpdateProcessName(String updateProcessName) {
-        this.updateProcessName = updateProcessName;
     }
 
     /**
